add trips model first part
