+++ Sources/OJP/OJP.swift
@@ -21,6 +21,7 @@ public enum LocationTypeRestriction: String {
 public class OJP {
     let loader: Loader
     let locationInformationRequest: OJPHelpers.LocationInformationRequest
+    let tripRequest: OJPHelpers.TripRequest
 
     /// Constructor of the OJP class
     /// - Parameter loadingStrategy: Pass a real loader with an API Configuration or a Mock for test purpuse
@@ -30,9 +31,11 @@ public class OJP {
             let httpLoader = HTTPLoader(configuration: apiConfiguration)
             loader = httpLoader.load(request:)
             locationInformationRequest = OJPHelpers.LocationInformationRequest(requesterReference: apiConfiguration.requesterReference)
+            tripRequest = OJPHelpers.TripRequest(requesterReference: apiConfiguration.requesterReference)
         case let .mock(loader):
             self.loader = loader
             locationInformationRequest = OJPHelpers.LocationInformationRequest(requesterReference: "Mock_Requestor_Ref")
+            tripRequest = OJPHelpers.TripRequest(requesterReference: "Mock_Requestor_Ref")
         }
     }
 
@@ -89,12 +92,27 @@ public class OJP {
 
         return locationInformationDelivery.placeResults
     }
+    
+    public func requestTrips(from originRef: String, destinationRef: String, viaRef: String? = nil) async throws -> [OJPv2.TripResult] {
+        let ojp = tripRequest.requestTrips(from: originRef, destinationRef: destinationRef, viaRef: viaRef)
+
+        let serviceDelivery = try await request(with: ojp).serviceDelivery
+
+        guard case let .trip(tripDelivery) = serviceDelivery.delivery else {
+            throw OJPSDKError.unexpectedEmpty
+        }
+
+        return tripDelivery.tripResults
+    }
+    
 
     func request(with ojp: OJPv2) async throws -> OJPv2.Response {
         let ojpXMLData = try encoder.encode(ojp, withRootKey: "OJP", rootAttributes: OJP.requestXMLRootAttributes)
-        guard String(data: ojpXMLData, encoding: .utf8) != nil else {
+        guard let xmlString = String(data: ojpXMLData, encoding: .utf8) else {
             throw OJPSDKError.encodingFailed
         }
+        
+        debugPrint(xmlString)
 
         let (data, response): (Data, URLResponse)
         do {
